+++ apple-flat-package/src/distribution.rs
@@ -14,7 +14,7 @@ use {
 };
 
 /// Represents a distribution XML file.
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename = "installer-gui-script", rename_all = "kebab-case")]
 pub struct Distribution {
     #[serde(rename = "minSpecVersion")]
@@ -59,18 +59,18 @@ impl Distribution {
     }
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct AllowedOsVersions {
     #[serde(rename = "os-version")]
     os_versions: Vec<OsVersion>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct App {
     pub id: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct Background {
     // TODO convert to enum.
@@ -82,7 +82,7 @@ pub struct Background {
     pub uti: Option<String>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Bundle {
     #[serde(rename = "CFBundleShortVersionString")]
     pub cf_bundle_short_version_string: Option<String>,
@@ -94,13 +94,13 @@ pub struct Bundle {
     // BuildVersion, SourceVersion reserved attributes.
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct BundleVersion {
     #[serde(default)]
     pub bundle: Vec<Bundle>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Choice {
     // The naming format on this element is all over the place.
     #[serde(rename = "customLocation")]
@@ -124,13 +124,13 @@ pub struct Choice {
     pub pkg_ref: Vec<PkgRef>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct ChoicesOutline {
     // ui is a reserved attribute.
     pub line: Vec<Line>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Conclusion {
     pub file: String,
     #[serde(rename = "mime-type")]
@@ -139,7 +139,7 @@ pub struct Conclusion {
     // language is a reserved attribute.
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Domains {
     pub enable_anywhere: bool,
     #[serde(rename = "enable_currentUserHome")]
@@ -148,7 +148,7 @@ pub struct Domains {
     pub enable_local_system: bool,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct InstallationCheck {
     pub script: Option<bool>,
     pub ram: Option<Ram>,
@@ -156,7 +156,7 @@ pub struct InstallationCheck {
     pub required_graphics: Option<RequiredGraphics>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct License {
     pub file: String,
@@ -165,25 +165,25 @@ pub struct License {
     // auto, language, and sla are reserved but not defined.
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Line {
     pub choice: String,
     #[serde(default, rename = "line")]
     pub lines: Vec<Line>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Locator {
     #[serde(rename = "search")]
     pub searches: Vec<Search>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct MustClose {
     pub app: Vec<App>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Options {
     #[serde(rename = "allow-external-scripts")]
     pub allow_external_scripts: Option<bool>,
@@ -199,13 +199,13 @@ pub struct Options {
 }
 
 /// Defines a range of supported OS versions.
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct OsVersion {
     pub before: Option<String>,
     pub min: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct PkgRef {
     pub active: Option<bool>,
     pub auth: Option<String>,
@@ -227,19 +227,19 @@ pub struct PkgRef {
     pub relocate: Vec<Relocate>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Product {
     pub id: String,
     pub version: Option<String>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Ram {
     #[serde(rename = "min-gb")]
     pub min_gb: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Readme {
     pub file: String,
     pub mime_type: Option<String>,
@@ -247,14 +247,14 @@ pub struct Readme {
     // language is reserved.
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Relocate {
     #[serde(rename = "search-id")]
     pub search_id: String,
     pub bundle: Bundle,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct RequiredBundles {
     pub all: Option<bool>,
     pub description: Option<String>,
@@ -262,19 +262,19 @@ pub struct RequiredBundles {
     pub bundles: Vec<Bundle>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct RequiredClDevice {
     #[serde(rename = "$value")]
     pub predicate: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct RequiredGlRenderer {
     #[serde(rename = "$value")]
     pub predicate: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct RequiredGraphics {
     pub description: Option<String>,
@@ -283,14 +283,14 @@ pub struct RequiredGraphics {
     pub required_gl_renderer: Option<RequiredGlRenderer>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Script {
     // language is a reserved attribute.
     #[serde(rename = "$value")]
     pub script: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub enum SearchValue {
     #[serde(rename = "bundle")]
     Bundle(Bundle),
@@ -298,7 +298,7 @@ pub enum SearchValue {
     Script(Script),
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct Search {
     pub id: String,
@@ -310,13 +310,13 @@ pub struct Search {
     pub value: SearchValue,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 pub struct Title {
     #[serde(rename = "$value")]
     pub title: String,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename = "kebab-case")]
 pub struct VolumeCheck {
     pub script: bool,
@@ -324,7 +324,7 @@ pub struct VolumeCheck {
     pub required_bundles: Option<RequiredBundles>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct Welcome {
     pub file: String,